# QuickStart/Fungi_api_swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fungi API",
    "description": "API for Fungi project",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development"
    }
  ],
  "paths": {
...
    "/Roles/GetRole": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get role",
        "description": "Receives information about the role by id",
        "operationId": "GetRole",
        "parameters": [
          {
            "name": "RoleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetRoleResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRoleResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRoleResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Roles/GetFilteredRoles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get filtered roles",
        "description": "Gets a list of roles with filters",
        "operationId": "GetFilteredRoles",
        "parameters": [
          {
            "name": "PartOfName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AccessLevelFrom",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "AccessLevelTo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetFilteredRolesResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFilteredRolesResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFilteredRolesResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Roles/CreateRole": {
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "Create role",
        "description": "Creates role",
        "operationId": "CreateRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRoleResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRoleResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRoleResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Roles/UpdateRole": {
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "Update role",
        "description": "Updates role info by id",
        "operationId": "UpdateRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateRoleResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateRoleResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateRoleResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Roles/DeleteRole": {
      "delete": {
        "tags": [
          "Roles"
        ],
        "summary": "Delete role",
        "description": "Deletes role by id",
        "operationId": "DeleteRole",
        "parameters": [
          {
            "name": "RoleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteRoleResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteRoleResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteRoleResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/Roles/GetRolePermissions": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get role permissions",
        "description": "Returns assigned permission codes for role",
        "operationId": "GetRolePermissions",
        "parameters": [
          {
            "name": "RoleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetRolePermissionsResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRolePermissionsResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRolePermissionsResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Roles/SetRolePermissions": {
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "Set role permissions",
        "description": "Replaces role permissions with provided permission codes",
        "operationId": "SetRolePermissions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRolePermissionsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRolePermissionsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetRolePermissionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SetRolePermissionsResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetRolePermissionsResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetRolePermissionsResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRoleRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "accessLevel": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreateRoleResponse": {
        "type": "object",
        "properties": {
          "createdRoleId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "CreateRoleResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateRoleResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
      "DeleteRoleResponse": {
        "type": "object",
        "properties": {
          "deletedRoleId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "DeleteRoleResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeleteRoleResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
      "ExceptionDto": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string"
          },
          "errorGroup": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "GetFilteredRolesResponse": {
        "type": "object",
        "properties": {
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            }
          }
        },
        "additionalProperties": false
      },
      "GetFilteredRolesResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GetFilteredRolesResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
      "GetRolePermissionsResponse": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "permissionCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "GetRolePermissionsResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GetRolePermissionsResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
      "GetRoleResponse": {
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/RoleDto"
          }
        },
        "additionalProperties": false
      },
      "GetRoleResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GetRoleResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
      "RoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "accessLevel": {
            "type": "integer",
            "format": "int32"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "SetRolePermissionsRequest": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "permissionCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "SetRolePermissionsResponse": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "permissionCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "SetRolePermissionsResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SetRolePermissionsResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateRoleRequest": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "newName": {
            "type": "string"
          },
          "newAccessLevel": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "UpdateRoleResponse": {
        "type": "object",
        "properties": {
          "updatedRoleId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "UpdateRoleResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UpdateRoleResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}